+++ src/OpenTelemetry.ResourceDetectors.Azure/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-* Added Azure Container Apps Detector to generate attributes: `service.name`,
-  `service.version`, `service.instance.id`, `cloud.provider` and
+* Added Azure Container Apps Resource Detector to generate attributes:
+  `service.name`, `service.version`, `service.instance.id`, `cloud.provider` and
   `cloud.platform`.
   ([#1565](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1565))